+++ src/app/core/models/index.ts
@@ -4,6 +4,5 @@ export * from './product-category.model';
 export * from './config-option.model';
 export * from './app-config.model';
 export * from './order.model';
-export * from './order-status.model';
 export  * from './app-settings.model';
 export * from './app-theme.model';