+++ src/app/core/models/order-status.model.ts
@@ -1,6 +1,0 @@
-export enum OrderStatus {
-    OPEN = 'OPEN',
-    SENT = 'SENT',
-    RECEIVED = 'RECEIVED',
-    CLOSED = 'CLOSED',
-}