+++ src/app/orders/orders.module.ts
@@ -1,8 +1,10 @@
 import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
 
 import { SharedModule } from 'src/app/shared/shared.module';
 import { OrderItemComponent, OrderListComponent, ProcessOrderComponent, } from './components';
 import { OrdersRoutingModule } from './orders-routing.module';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 
 @NgModule({
     declarations: [
@@ -10,7 +12,7 @@ import { OrdersRoutingModule } from './orders-routing.module';
         OrderListComponent,
         OrderItemComponent,
     ],
-    imports: [SharedModule, OrdersRoutingModule],
+    imports: [SharedModule, OrdersRoutingModule, ReactiveFormsModule, MatCheckboxModule],
     exports: [OrderListComponent, OrdersRoutingModule],
 })
 export class OrdersModule {