+++ src/app/shared/validators/validators.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AsyncEmailValidatorDirective } from './async-email-validator.directive';
+
+@NgModule({
+    declarations: [AsyncEmailValidatorDirective],
+    imports: [
+        CommonModule
+    ],
+    exports: [AsyncEmailValidatorDirective]
+})
+export class ValidatorsModule {
+}